25/08/2023 commit

+++ C0523G1-duydd-co523g1_Module2.1/furama_resort/src/repository/people/EmployeeRepositoryImpl.java
@@ -18,7 +18,8 @@ public class EmployeeRepositoryImpl implements IEmployeeRepository {
         String[] arr = null;
         for (String s : strings) {
             arr = s.split(",");
-            employees.add(new Employee(arr[0], arr[1], arr[2], arr[3], arr[4], arr[5], arr[6], arr[7], arr[8], Integer.parseInt(arr[9])));
+            employees.add(new Employee(arr[0], arr[1], arr[2], arr[3], arr[4]
+                    , arr[5], arr[6], arr[7], arr[8], Double.parseDouble(arr[9])));
         }
         return employees;
     }
